Extracted ISymbol extension methods to separate file
